bot/requirements.txt
azure-functions>=1.18.0
azure-communication-callautomation>=1.1.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-appconfiguration>=1.5.0
msgraph-sdk>=1.0.0
rapidfuzz>=3.6.0
jellyfish>=1.0.0
azure-monitor-opentelemetry>=1.2.0
